General statistical tools

Running mean and variance

There are plenty of implementations of mean and variance of numbers stored in a vector. However, sometimes it is more convenient to maintain running estimates of mean and/or variance, adding new data as it pops up, without storing the whole dataset in an array. One way of achieving this is with West's recursion formula, which we implement here.

Unweighted data

BioStatPhys.MeanVar — Type

MeanVar

Type to allow running computation of mean and variance (i.e. without storing the data points).

After the object is created with MeanVar(), data are fed one at a time with push!, causing the internal state to be updated but without storing the new data point (MeanVar requires a fixed amount of storage: two real numbers and an integer). Mean and variance estimates from a MeanVar object are obtained calling mean and var.

MeanVar uses West's recursion formula, see

D. H. D. West, Communications of the ACM 22, 532 (1979)

Example

mv=MeanVar()
push!(mv,a)
push!(mv,b)
...
println("Mean = $(mean(mv))")
println("Variance = $(var(mv))")

See also

See WMeanVar for the case of weighted data.

Base.push! — Method

function push!(MV::MeanVar,x)

Add data point x to MeanVar object MV

BioStatPhys.mean — Method

Compute mean estimate (sample mean) from a MeanVar object

BioStatPhys.var — Method

Compute variance estimate (population variance) from a MeanVar object

Weighted data

BioStatPhys.WMeanVar — Type

WMeanVar

Type to allow running computation of mean and variance of weighted data (without storing the data points).

The object is created with WMeanVar(), data are fed one at a time with push!, and mean and variance are obtained calling mean and var.

Interface is the same as for the unweighted version `MeanVar', which see for an example.

WMeanVar uses West's recursion formula, see

D. H. D. West, Communications of the ACM 22, 532 (1979)

see MeanVar

Base.push! — Method

function push!(WMV::WMeanVar,x)

Add data point x with weight W to WMeanVar object WMV

BioStatPhys.mean — Method

Compute mean estimate (sample mean) from a WMeanVar object

BioStatPhys.var — Method

Compute variance estimate (population variance) from a WMeanVar object