Miscellaneous tools

Real-indexed vector

This is a simple implementation of a vector that can be accessed with real (floating point) sub-indices. It does simply a linear binning of a specified (semi-open) interval $[min,max)$ on the real line, and provides a convenient Vector-like syntax to access elements. The stored values can be of any type, but the interval and number of bins are fixed at construction.

julia> 	A = BinnedVector{String}(5,min=0.,max=1.)5-element BinnedVector{String}:
 #undef
 #undef
 #undef
 #undef
 #undef
julia> 	A[0.1]="hello""hello"
julia> 	A[0.2]="bye bye""bye bye"
julia> 	A5-element BinnedVector{String}:
    "hello"
    "bye bye"
 #undef
 #undef
 #undef

To initialise elements with a default value, give function with a signature like zeros:

B = BinnedVector{Rational{Int64}}(10,min=2.,max=8.,init=zeros)
B[2.1]=3//5
B[5.3]=4//1
B

10-element BinnedVector{Rational{Int64}}:
 3//5
 0//1
 0//1
 0//1
 0//1
 4//1
 0//1
 0//1
 0//1
 0//1

Indexing with integers refers to bin numbers.

julia> 	A[1:2]2-element Vector{String}:
 "hello"
 "bye bye"
julia> 	B[1:5]5-element Vector{Rational{Int64}}:
 3//5
 0//1
 0//1
 0//1
 0//1

Outliers are mapped to bin 0 if below range, or -1 if above range.

julia> B[0.]=10//3 ; B[40.]=99//2;
julia> B[0]10//3
julia> B[-1]99//2

The Histogram type is implemented using BinnedVector.

API

BioStatPhys.BinnedVector — Type

mutable struct BinnedVector{T}

Simple real-indexed vector. Maps a real interval to an integer range 1:nbins then used to index a vector of arbitrary type. Can be used e.g. to build histograms. The real interval and number of bins nbins are fixed at the outset.

Example

Create with

A = BinnedVector{Int}(nbins,min=1.,max=10.,init=zeros)

(init is optional, defaults to leave elements undefined), and access or write as a vector:

A[5.2] += 2

Numbers above and below range map to two special bins.

If indexed with integers, these are interpreted as bin numbers. A[0] and A[-1] are the outlier bins (below and above, respectively).

BioStatPhys.interval — Function

interval(A::BinnedVector{T})

Return tuple (min,max) giving the extrema of the real interval mapped to the array bins.

BioStatPhys.delta — Function

delta(A::BinnedVector{T})

Return the with of the bins of A

BioStatPhys.nbins — Function

nbins(A::BinnedVector{T})

Return number of bins of A

BioStatPhys.bin — Function

bin(A::BinnedVector{T},x::Float64) where {T}

Map real value x to bin number. Return 0 if below range, or -1 if above range.

BioStatPhys.binc — Function

binc(A::BinnedVector{T},i::Int) where {T}

Return center of bin i, which must be in range 1:size(A,1). Does not perform range check.